Added userled to arduino-r4-minima

+++ arch/arm/src/ra/ra_gpio.h
@@ -57,8 +57,8 @@ extern "C"
 
 typedef struct gpio_pinset
 {
-    uint32_t port;
-    uint32_t pin;
+    uint8_t port;
+    uint8_t pin;
     uint32_t cfg;
 }gpio_pinset_t;
